refactoring SmallFilmCard component
