+++ project/src/components/small-film-card/SmallFilmCard.tsx
